FIX(#004): CHANGE LONG WORDS DISPLAY IN COURSE CLASSES
